// assets/js/enum.ts
export enum TYPE {
    NATION = 'N',
    CITY = 'C',
    DISC = 'D',
    VLI = 'L',
}

export enum PARTY {
    KMT = "中國國民黨",
    DPP = "民主進步黨",
    TPP = "台灣民眾黨",
    NP = "新黨",
    AAA = "親民黨",
    OTHER = "",
}

export enum COLOR {
    KMT = '#8894D8',
    DPP = '#6AA27A',
    TPP = '#84CB98',
    NP = '#DFA175',
    OTHER = '#BFBFBF',
}